+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 [Full changelog](https://github.com/mozilla-rally/core-addon/compare/v1.2.1...master)
 
+* [#555](https://github.com/mozilla-rally/rally-core-addon/pull/555): Update the income question in the demographic survey to take in an integer.
 * [#601](https://github.com/mozilla-rally/rally-core-addon/pull/601): Enable submitting data with Glean in addition to the legacy collection.
 
 # v1.2.1 (2021-04-27)